+++ index.ts
@@ -23,6 +23,7 @@ export function register() {
     R.Config("unlock_all_songs", { type: "boolean", default: false })
     R.Config("unlock_all_items", { type: "boolean", default: false })
     R.Config("comment_feature", { type: "boolean", default: true })
+    R.Config("<groovin'_upper>_fill_upper_points", { type: "boolean", default: false })
 
     R.WebUIEvent("removeWebUIMessage", UtilityHandlersWebUI.removeWebUIMessage)